Fix compiler warnings and documentation issues
- Add compiler flags to suppress unused parameter/variable warnings
- Fix pointer vs character comparison bug in options.c
- Fix size_t comparison warning in stream.c
- Correct Texinfo node navigation to match menu order

+++ stream.c
@@ -329,7 +329,7 @@ int TreeCCStreamFlushStdio(TreeCCStream *stream, FILE *file)
 	{
 		if(buffer == stream->lastBuf)
 		{
-			if(fwrite(buffer->data, 1, stream->posn, file) != stream->posn)
+			if(fwrite(buffer->data, 1, stream->posn, file) != (size_t)stream->posn)
 			{
 				return 0;
 			}
